Video = true, audio = false fix
Extremely long winded in the way it’s coded - if I was smart, I
would’ve created a helper function to handle all of the different
cases. Oh well.

Basically, this is the final chunk of code for handling the following
scenarios when calling another user:
- Both Users have a microphone and camera
- One or both users are missing either one of a microphone or camera
- One or both users are missing both a camera and mic
- or any combination of any of these scenarios.

Still some issues where if a user who is missing a mic calls someone
who HAS a mic, volume won’t work for either user, but if the user who
has a mic calls then it works fine. Kind of confused on why stuff like
that happens, but whatcha gonna do.